new update

+++ database/migrations/2024_01_01_000023_create_esi_records_table.php
@@ -22,7 +22,7 @@ return new class extends Migration {
             $table->string('member_name')->nullable();
             $table->string('dispensary')->nullable();
             $table->decimal('gross_wage', 15, 2)->default(0);
-            $table->integer('days_worked')->nullable();
+            $table->decimal('days_worked', 5, 1)->nullable();
             $table->string('ee_0_75')->nullable();
             $table->string('er_3_25')->nullable();
             $table->decimal('total_contribution', 15, 2)->default(0);
